requirements.txt
ccxt>=4.2.0
python-dotenv>=1.0.0
textual>=0.47.0
pydantic>=2.5.0
asyncio>=3.4.3
websockets>=12.0
aiohttp>=3.9.0
